[not verified] rename to ConnectedProductOffer
