+++ projects/packages/my-jetpack/_inc/components/connected-product-offer/stories/mock-data.js
