+++ projects/packages/my-jetpack/_inc/components/product-cards-section/stories/index.jsx
@@ -4,7 +4,7 @@
 import React from 'react';
 import withMock from 'storybook-addon-mock';
 import { HashRouter, Routes, Route } from 'react-router-dom';
-import { getAllMockData } from '../../connected-product-detail/stories/utils.js';
+import { getAllMockData } from '../../connected-product-offer/stories/utils.js';
 
 /**
  * Internal dependencies
