+++ projects/packages/my-jetpack/_inc/components/product-interstitial/stories/index.jsx
@@ -9,7 +9,7 @@ import { HashRouter, Routes, Route } from 'react-router-dom';
 /**
  * Internal dependencies
  */
-import { getMockData } from '../../connected-product-detail/stories/utils.js';
+import { getMockData } from '../../connected-product-offer/stories/utils.js';
 import ProductInterstitial, {
 	AntiSpamInterstitial,
 	BackupInterstitial,
