--- projects/packages/my-jetpack/changelog/update-my-jetpack-add-connected-product-card
@@ -1,4 +1,0 @@
-Significance: patch
-Type: added
-
-My Jetpack: add ConnectedProductDetail component
