+++ projects/packages/my-jetpack/changelog/update-my-jetpack-add-connected-product-offer
@@ -1,0 +1,4 @@
+Significance: patch
+Type: added
+
+My Jetpack: add ConnetedProductOffer component
